+++ src/scores/continuous/__init__.py
@@ -6,7 +6,6 @@ from scores.continuous.flip_flop_impl import (
     flip_flop_index,
     flip_flop_index_proportion_exceeding,
 )
-from scores.continuous.isoreg_impl import isotonic_fit
 from scores.continuous.murphy_impl import murphy_score, murphy_thetas
 from scores.continuous.quantile_loss_impl import quantile_score
 from scores.continuous.standard_impl import (
@@ -21,7 +20,6 @@ from scores.continuous.standard_impl import (
 __all__ = [
     "flip_flop_index",
     "flip_flop_index_proportion_exceeding",
-    "isotonic_fit",
     "murphy_score",
     "murphy_thetas",
     "quantile_score",
