+++ src/scores/probability/__init__.py
@@ -2,7 +2,6 @@
 Import the functions from the implementations into the public API
 """
 
-from scores.continuous.isoreg_impl import isotonic_fit
 from scores.continuous.murphy_impl import murphy_score, murphy_thetas
 from scores.probability.brier_impl import brier_score
 from scores.probability.crps_impl import (
@@ -14,7 +13,6 @@ from scores.probability.crps_impl import (
 from scores.probability.roc_impl import roc_curve_data
 
 __all__ = [
-    "isotonic_fit",
     "murphy_score",
     "murphy_thetas",
     "brier_score",
