+++ src/scores/processing/__init__.py
@@ -8,6 +8,7 @@ from scores.processing.discretise import (
     comparative_discretise,
     proportion_exceeding,
 )
+from scores.processing.isoreg_impl import isotonic_fit
 from scores.processing.matching import broadcast_and_match_nan
 
 __all__ = [
@@ -16,4 +17,5 @@ __all__ = [
     "proportion_exceeding",
     "broadcast_and_match_nan",
     "binary_discretise_proportion",
+    "isotonic_fit",
 ]
